tested the add and remove methods
