+++ package.json
@@ -275,7 +275,6 @@
     "has-tostringtag": "^1.0.0",
     "has-yarn": "^2.1.0",
     "he": "^1.2.0",
-    "hint": "^7.1.1",
     "hosted-git-info": "^5.1.0",
     "hpack.js": "^2.1.6",
     "html-encoding-sniffer": "^3.0.0",
@@ -736,15 +735,22 @@
     "zwitch": "^1.0.5"
   },
   "devDependencies": {
+    "@babel/core": "^7.18.13",
+    "@babel/preset-env": "^7.18.10",
+    "babel-jest": "^29.0.1",
     "css-loader": "^6.7.1",
+    "hint": "^7.1.1",
     "html-loader": "^4.1.0",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.0.1",
+    "jest-localstorage-mock": "^2.4.22",
     "style-loader": "^3.3.1",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.10.1"
   },
   "scripts": {
+    "test": "jest",
     "start": "webpack serve --open",
     "deploy": "gh-pages -d dist",
     "build": "webpack --config webpack.config.js"
